pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workplace-ai-env"
version = "1.0.0"
description = "Multi-Agent Decision Intelligence Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "WorkplaceAIEnv Team"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "huggingface-hub>=0.20.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["workplace_ai_env*"]

[tool.pytest.ini_options]
testpaths = ["workplace_ai_env/tests"]
asyncio_mode = "auto"
